+++ libs/web/src/grids/group-permissions-grid/group-permissions-grid.component.ts
@@ -89,7 +89,7 @@ export class GroupPermissionsGridComponent extends BaseResourcesGridComponent {
     const itemModal: PermissionModalComponent = this.app.modals(this.resolver).create(PermissionModalComponent);
     itemModal.name = 'editGroupPermission';
     itemModal.account = this.accountService.account;
-    itemModal.readonly = !this.account.checkPermissions(['change_permission']) || this.readonly;
+    itemModal.readonly = !this.checkPermissions(['change_permission']) || this.readonly;
     itemModal.text = this.translateService.instant('Save');
     itemModal.title = this.translateService.instant('Edit permission');
     if (itemModal.readonly) {