+++ libs/web/src/pages/components-page/controls-frame/controls-frame.component.ts
@@ -29,6 +29,6 @@ export class ControlsFrameComponent extends BaseFrameComponent {
     super(accountService, app, translateService, activatedRoute, router, sharedService);
   }
   get readonly() {
-    return !this.account || !this.account.checkPermissions(['change_controls']);
+    return !this.checkPermissions(['change_controls']);
   }
 }