+++ typings.d.ts
@@ -1,6 +1,8 @@
-// Typings reference file, you can add your own global typings here
-// https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-
+/* SystemJS module definition */
+declare var module: NodeModule;
+interface NodeModule {
+  id: string;
+}
 // tslint:disable
 declare const System: any;
 declare var node: any;